+++ package.json
@@ -21,6 +21,7 @@
     "eslint": "8.40.0",
     "eslint-config-next": "13.4.1",
     "firebase": "^10.1.0",
+    "firebase-admin": "^11.10.1",
     "next": "13.4.1",
     "react": "18.2.0",
     "react-chartjs-2": "^5.2.0",
